Recv stuff

+++ src/networking/NetManager.hpp
@@ -81,7 +81,7 @@ class NetManager {
         // Receive data and then parse it
         void receiveData();
         // Parse data. Called in receiveData.
-        Result<int> parseData(const CTSerialize::MessageHeader* msg);
+        Result<uint8_t> parseData(const CTSerialize::MessageHeader* msg);
 
         // Steamnetworking has no way to kick/ban users. I think? Instead just remove the dudes access to m_playersInLobby and any
         // Further attempts to join (this member is specificaly for fetchMemberList())
